requirements.txt
djangorestframework
Django 
git+https://github.com/DUB1401/dublib#egg=dublib
PyJWT
